app/requirements.txt
flask
flask-login
flask_sqlalchemy
flask_bcrypt
Flask-Migrate
influxdb_client
requests
pandas
psycopg2
psycopg2-binary
python-dotenv
python-dateutil
pyyaml
numpy